Separated projects using service collection extensions
