+++ Portfolio.API/Controllers/SentimentAnalysisController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.ML;
-using Portfolio.Data;
 using Portfolio.Logic.Interfaces;
+using Portfolio.ML_Models.Data;
 using System;
 
 namespace Portfolio.API.Controllers
