+++ Portfolio.ML_Models/Data/SentimentPrediction.cs
@@ -1,25 +1,15 @@
 ﻿using Microsoft.ML.Data;
+using System;
+using System.Collections.Generic;
+using System.Text;
 
-namespace Portfolio.Data
+namespace Portfolio.ML_Models.Data
 {
-    public class SentimentData
-    {
-        [LoadColumn(0)]
-        public string SentimentText;
-
-        [LoadColumn(1)]
-        [ColumnName("Label")]
-        public bool Sentiment;
-    }
-
     public class SentimentPrediction : SentimentData
     {
-
         [ColumnName("PredictedLabel")]
         public bool Prediction { get; set; }
-
         public float Probability { get; set; }
-
         public float Score { get; set; }
     }
 }
